refactor: account creation
