+++ ui/config/messages/en.json
@@ -130,7 +130,7 @@
       "img": "/images/applets/account-creation.svg",
       "title": "Subaccount creation",
       "description": "Create subaccounts.",
-      "path": "/create-account",
+      "path": "/account/create",
       "keywords": "account,create,register"
     },
     "5": {
