--- ui/portal/web/src/components/create-account/index.ts
@@ -1,3 +1,0 @@
-export { CreateAccountWrapper } from "./CreateAccountWrapper";
-export { CreateAccountTypeStep } from "./CreateAccountTypeStep";
-export { CreateAccountDepositStep } from "./CreateAccountDepositStep";
